+++ webv2/primarykey/primarykey_test.go
@@ -1,4 +1,21 @@
-package webv2
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package web defines web APIs to be used with harbourbridge frontend.
+// Apart from schema conversion, this package involves API to update
+// converted schema.
+package primarykey
 
 import (
 	"bytes"
@@ -56,7 +73,7 @@ func TestUpdatePrimaryKey(t *testing.T) {
 	req.Header.Set("Content-Type", "application/json")
 	rr := httptest.NewRecorder()
 
-	handler := http.HandlerFunc(primaryKey)
+	handler := http.HandlerFunc(PrimaryKey)
 	handler.ServeHTTP(rr, req)
 
 	var res PrimaryKeyResponse
@@ -116,7 +133,7 @@ func TestAddPrimaryKey(t *testing.T) {
 	req.Header.Set("Content-Type", "application/json")
 	rr := httptest.NewRecorder()
 
-	handler := http.HandlerFunc(primaryKey)
+	handler := http.HandlerFunc(PrimaryKey)
 	handler.ServeHTTP(rr, req)
 
 	var res PrimaryKeyResponse
@@ -181,7 +198,7 @@ func TestRemovePrimaryKey(t *testing.T) {
 	req.Header.Set("Content-Type", "application/json")
 	rr := httptest.NewRecorder()
 
-	handler := http.HandlerFunc(primaryKey)
+	handler := http.HandlerFunc(PrimaryKey)
 	handler.ServeHTTP(rr, req)
 
 	var res PrimaryKeyResponse
@@ -268,7 +285,7 @@ func TestPrimarykey(t *testing.T) {
 		req.Header.Set("Content-Type", "application/json")
 		rr := httptest.NewRecorder()
 
-		handler := http.HandlerFunc(primaryKey)
+		handler := http.HandlerFunc(PrimaryKey)
 		handler.ServeHTTP(rr, req)
 
 		json.Unmarshal(rr.Body.Bytes(), &tt.res)
